Split MDev off into a separate module
